03-FaceRecognition-I/thetensorclan-aws/requirements.txt
https://download.pytorch.org/whl/cpu/torch-1.6.0%2Bcpu-cp37-cp37m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.7.0%2Bcpu-cp37-cp37m-linux_x86_64.whl
boto3==1.9.66
Flask==1.1.2
dlib==19.21.0
numpy==1.19.1
Pillow==7.2.0
opencv-python==4.4.0.40